4.21*: bloglist expansion, step9
What:

1. Change the delete blog operation so that a blog can be deleted only by the user who added the blog.
2. If deleting a blog is attempted without a token or by an invalid user, the operation should return 403 code.

+++ part4/blog-list-backend/controllers/blogs.js
@@ -32,7 +32,18 @@ blogsRouter.post('', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = await jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if(decodedToken.id !== blog.user.toString()){
+    return response.status(403).json({ error: 'invalid user' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
+
   response.status(204).end()
 })
